chore: rename files name that might cause ci fail
