+++ frontend/rust-lib/event-integration/tests/document/af_cloud_test/edit_test.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-use event_integration::test_document::assert_document_data_equal;
+use event_integration::document_event::assert_document_data_equal;
 use flowy_document2::entities::DocumentSyncStatePB;
 
 use crate::document::af_cloud_test::util::AFCloudDocumentTest;
